Part 1/2 *Applying shady's AI & Spells patch, Requires *SQL update* http://www.aspiredev.org/forums/viewtopic.php?f=22&t=24557 AI: AI_AGENTS cooldown loading improved. AI: Spellcasting AI improved. AI: Mobs shouldn't regen mana ignoring 5 seconds rule. (Also mana and focus regeneration was strongly nerfed) AI: Mobs under fear effects will not regen HP and will not wipe their aggro list. AI: Fixed loads problems with neutrals. AI: Fixed aggro gain to mobs under stuns, fear etc. Stability: crashfix with Mechanics (btw assertion to constants sucks a lot! ) Spells: Cloak of shadows should not break from stealth. Spells: all triggered spells should NOT break stealth! (bye bye hackfixes) Spells: Improved Vampiric Embrace rewritten. Fully working, faster and more clean. (but still a hackfix) Spells: Vampiric Touch implementation removed since it's out of date (VT in WotLK works in different way) Spells: Shadowguard spellfixes removed since there is no such spell any more. Spells: Arcane Blast will cast debuff poperly Spells: Fear Ward shouldn't be able to cast while in fear Spells: Vanish shouldn't remove stealth over 10 seconds.
git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@1245 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Player.h
@@ -2080,8 +2080,8 @@ public:
 	uint32 m_outStealthDamageBonusPeriod;
 	uint32 m_outStealthDamageBonusTimer;
 
-	int16 m_vampiricEmbrace;
-	int16 m_vampiricTouch;
+	uint32 m_vampiricEmbrace;
+	uint32 m_vampiricTouch;
 	void VampiricSpell(uint32 dmg, UnitPointer pTarget);
 
 	// grounding totem
